Refactor owners.mjs into it's own script
