+++ .github/scripts/owners-review-recommendations.mjs
@@ -1,137 +1,7 @@
-import { readFileSync } from "node:fs";
-import { join } from "node:path";
-import { minimatch } from "minimatch";
-import { ensureEnvVar, initGithub } from "./github-helpers.mjs";
+import { ensureEnvVar, getChangedFiles, initGithub } from "./github-helpers.mjs";
+import { assignOwnership, ownershipsToAllocations, parseOwnersFile } from "./owners.mjs";
 
-// Resolve relative to this file so the path works regardless of cwd
-// (workflow runs from repo root; `npm test` runs from .github/scripts).
-const OWNERS_FILE = join(import.meta.dirname, "..", "OWNERS");
-
-/**
- * @typedef Owner
- * @property { string } filepath
- * @property { string } team
- * */
-
-/**
- * @typedef Allocation
- * @property { string } team
- * @property { number } fileCount
- * */
-
-/**
- * @typedef { Map<string, string[]> } Ownerships
- * */
-
-/**
- * @returns { Owner[] }
- * */
-export function parseOwnersFile() {
-	const content = readFileSync(OWNERS_FILE, "utf8");
-
-	const owners = content.split("\n")
-	.filter(line => line.includes("@n8n-io"))
-	.map(line => ({
-			filepath: line.match(/^\S+/)?.at(0),
-			team: line.match(/@n8n-io\/.*/)?.at(0)
-		}))
-	.filter(/** @returns { owner is Owner } */ (owner) =>
-		owner.filepath !== undefined && owner.team !== undefined
-	);
-
-	return owners;
-}
-
-/**
- * Translate a CODEOWNERS-style pattern into a minimatch glob.
- *   "*"            -> "**"        (catch-all)
- *   "packages/x/"  -> "packages/x/**"  (directory, recursive)
- *   "path/to/f.ts" -> "path/to/f.ts"   (exact file or already a glob)
- *
- * @param { string } pattern
- * @returns { string }
- * */
-function codeownersToMinimatch(pattern) {
-	if (pattern === "*") return "**";
-	if (pattern.endsWith("/")) return pattern + "**";
-	return pattern;
-}
-
-/**
- * Map each changed file to the team that owns it, applying CODEOWNERS
- * last-match-wins semantics. Files that match no rule are omitted.
- *
- * @param { Set<string> } files
- * @param { Owner[] } owners
- *
- * @returns { Ownerships } team -> files it owns in this changeset
- * */
-export function assignOwnership(files, owners) {
-	const compiled = owners.map(owner => ({
-		team: owner.team,
-		glob: codeownersToMinimatch(owner.filepath),
-	}));
-
-	/** @type { Map<string, string[]> } */
-	const teamToFiles = new Map();
-
-	for (const file of files) {
-		// Walk rules in reverse so the *last* matching rule wins.
-		for (let i = compiled.length - 1; i >= 0; i--) {
-			if (minimatch(file, compiled[i].glob, { dot: true })) {
-				const team = compiled[i].team;
-				const bucket = teamToFiles.get(team);
-
-				if (bucket) {
-					bucket.push(file);
-				} else {
-					teamToFiles.set(team, [file]);
-				}
-				break;
-			}
-		}
-	}
-
-	return teamToFiles;
-}
-
-/**
- * @param { number } pullRequestNumber
- *
- * @returns { Promise<Set<string>> }
- * */
-export async function getChangedFiles(pullRequestNumber) {
-	const { octokit, owner, repo, } = initGithub();
-
-	const files = await octokit.paginate(
-		octokit.rest.pulls.listFiles,
-		{
-			owner,
-			repo,
-			pull_number: pullRequestNumber,
-			per_page: 100
-		}
-	);
-
-	const filenames = new Set([
-		...files.map(file => file.filename),
-		...files.map(file => file.previous_filename).filter(filename => filename !== undefined)
-	]);
-
-	return filenames;
-}
-
-/**
- * @param { Ownerships } ownerships
- *
- * @returns { Allocation[] }
- * */
-export function ownershipsToAllocations(ownerships) {
-	return Array.from(ownerships).map(([team, files]) => ({
-			team,
-			fileCount: files.length
-		}))
-}
+/** @typedef {import('./owners.mjs').Allocation} Allocation */
 
 /**
  * @param { number } pullRequestNumber
@@ -143,7 +13,7 @@ export async function getReviewRecommendations(pullRequestNumber) {
 	const ownerships = assignOwnership(changedFiles, owners);
 	const allocations = ownershipsToAllocations(ownerships);
 
-	const topAllocations = allocations.sort((a,b) => b.fileCount - a.fileCount).slice(0, 3);
+	const topAllocations = allocations.sort((a, b) => b.fileCount - a.fileCount).slice(0, 3);
 
 	await commentOnPrWithRecommendations(pullRequestNumber, topAllocations, changedFiles);
 }
